Изменить обработку RequestBody в /change-quantity, shopping-cart

+++ commerce/shopping-cart/src/main/java/ru/yandex/practicum/request/ChangeProductQuantityRequest.java
@@ -13,5 +13,5 @@ import java.util.UUID;
 public class ChangeProductQuantityRequest {
     UUID productId;
 
-    Integer newQuantity;
+    Long newQuantity;
 }
